requirements.txt
albumentations==1.3.0
dash==2.9.1
einops==0.6.1
ipympl==0.9.3
ipywidgets==8.0.5
jupyter-dash==0.4.2
kaleido==0.2.1
kornia==0.6.12
lion-pytorch==0.0.7
matplotlib==3.7.1
nbformat==5.8.0
numpy==1.23.5
omegaconf==2.3.0
opencv-python==4.7.0.72
pandas==1.5.3
pillow==9.4.0
plotly==5.14.1
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
pytorch-lightning==2.0.1.post0
rich==13.3.4
scipy==1.10.1
seaborn==0.12.2
timm==0.6.13
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
umap-learn==0.5.3
wandb==0.15.3
